+++ ishtech-spring-boot-jwt-web/src/main/resources/db/migration/V20251021_020418__create_table_user_role.sql
@@ -1,6 +1,6 @@
 CREATE TABLE t_user_role (
-  id        SERIAL       PRIMARY KEY,
-  user_id   INT          NOT NULL,
+  id        BIGSERIAL    PRIMARY KEY,
+  user_id   BIGINT       NOT NULL,
   role_name VARCHAR(255) NOT NULL,
   is_active BOOLEAN      NOT NULL DEFAULT TRUE,
   CONSTRAINT uk_user_role_user_id_role UNIQUE (user_id, role_name),
