+++ ishtech-spring-boot-jwt-web/src/main/resources/db/migration/V20251021_020450__create_table_user_role_aud.sql
@@ -1,9 +1,9 @@
 CREATE TABLE ${audit_schema_name}.t_user_role_aud (
-  id            INT           NOT NULL,
-  rev           INT           NOT NULL,
+  id            BIGINT        NOT NULL,
+  rev           BIGINT        NOT NULL,
   revtype       SMALLINT      NOT NULL,
   rev_timestamp TIMESTAMP     NOT NULL DEFAULT CURRENT_TIMESTAMP,
-  user_id   INT                   NULL,
+  user_id   BIGINT                NULL,
   role_name VARCHAR(255)          NULL,
   is_active BOOLEAN               NULL,
   PRIMARY KEY(id, rev)
